third_party/WebKit/LayoutTests/fast/css/getComputedStyle/computed-style-font-family-monospace-expected.txt

Test that only exactly 'font-family: monospace;' causes use of fixed-width font default size. All other font family and font family combinations should use the standard default size.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS fontSizeForFamilies('monospace') is "13px"
PASS fontSizeForFamilies('monospace, times') is "16px"
PASS fontSizeForFamilies('times, monospace') is "16px"
PASS fontSizeForFamilies('foo') is "16px"
PASS fontSizeForFamilies('foo, monospace') is "16px"
PASS successfullyParsed is true

TEST COMPLETE
